data upload on mongodb and cloudinary

+++ src/utils/cloudinary.utils.js
@@ -17,10 +17,19 @@ const uploadOnCloudinary = async (localFilePath) => {
       resource_type: "auto",
     });
     //file has been uploaded successfully
-    console.log("File is uploaded on cloudinary", response.url);
+    // console.log("File is uploaded on cloudinary", response.url);
+
+    // to unlink file if uploaded successfully
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath); // remove the locally saved temporary
+    }
+
     return response;
+    
   } catch (error) {
-    fs.unlinkSync(localFilePath); // remove the locally saved temporary
+    if (fs.existsSync(localFilePath)) {
+      fs.unlinkSync(localFilePath);
+    } // remove the locally saved temporary
     // file as the upload operation got failed
     return null;
   }
